+++ insh/src/components/insh.rs
@@ -4,16 +4,14 @@ use crate::components::file_creator::{
 };
 use crate::components::finder::{Finder, FinderEffect, FinderProps};
 use crate::components::searcher::{Searcher, SearcherEffect, SearcherProps};
-use crate::config::{BrowserSortHiddenConfig, Config};
+use crate::config::Config;
 use crate::current_dir;
 use crate::programs::{Bash, Vim};
 use crate::stateful::Stateful;
+use crate::utils::get_files_request;
 
 use file_type::FileType;
-use insh_api::{
-    FileSortOptions, FindFilesRequestParams, GetFilesRequestParams, HiddenFileSort, Request,
-    RequestParams, Response,
-};
+use insh_api::{FindFilesRequestParams, Request, RequestParams, Response};
 use rend::{Fabric, Size};
 use term::{Key, KeyEvent, KeyMods, TermEvent};
 use til::{Component, Event, SystemEffect};
@@ -321,25 +319,8 @@ impl From<Props> for State {
 impl State {
     fn browse(&mut self, dir: PathBuf, file: Option<PathBuf>) -> Option<SystemEffect<Request>> {
         // Create a request for getting the files in the dir.
-        let sort: Option<FileSortOptions> = self.config.browser().sort().map(|sort| {
-            FileSortOptions::builder()
-                .case_insensitive(sort.case_insensitive())
-                .hidden(match sort.hidden() {
-                    BrowserSortHiddenConfig::First => HiddenFileSort::First,
-                    BrowserSortHiddenConfig::Last => HiddenFileSort::Last,
-                    BrowserSortHiddenConfig::Mixed => HiddenFileSort::Mixed,
-                })
-                .build()
-        });
-
-        let request = Request::builder()
-            .params(RequestParams::GetFiles(
-                GetFilesRequestParams::builder()
-                    .dir(dir.clone())
-                    .sort(sort)
-                    .build(),
-            ))
-            .build();
+        let request =
+            get_files_request(dir.clone(), &self.config, self.config.browser().metadata());
 
         self.mode = Mode::Browse;
         let size: Size = Size::from(terminal::size().unwrap());