+++ insh/src/main.rs
@@ -30,6 +30,7 @@ mod requester;
 mod response_handler;
 mod stateful;
 mod string;
+mod utils;
 
 use std::os::unix::net::UnixStream;
 use std::process::exit;
@@ -40,20 +41,19 @@ use flexi_logger::LoggerHandle;
 use uuid::Uuid;
 
 use common::paths::INSHD_SOCKET;
-use insh_api::{
-    FileSortOptions, GetFilesRequestParams, HiddenFileSort, Request, RequestParams, Response,
-};
+use insh_api::{Request, Response};
 use term::TermEvent;
 use til::{App, AppRunOptions, Component, Requester, ResponseHandler, Stopper, SystemEffect};
 
 use crate::args::Args;
 use crate::components::{Insh, InshProps};
-use crate::config::{BrowserSortHiddenConfig, Config};
+use crate::config::Config;
 #[cfg(feature = "logging")]
 use crate::logging::{configure_logging, ConfigureLoggingResult};
 use crate::requester::InshdRequester;
 use crate::response_handler::{InshdResponseHandler, InshdResponseHandlerStopper};
 use crate::stateful::Stateful;
+use crate::utils::get_files_request;
 
 fn main() {
     let args: Args = Args::parse();
@@ -84,25 +84,11 @@ fn main() {
     // Determine the starting effects.
     let mut starting_effects: Option<Vec<SystemEffect<Request>>> = args.starting_effects();
     let pending_browser_request: Option<Uuid> = if args.browse() {
-        let sort: Option<FileSortOptions> = config.browser().sort().map(|sort| {
-            FileSortOptions::builder()
-                .case_insensitive(sort.case_insensitive())
-                .hidden(match sort.hidden() {
-                    BrowserSortHiddenConfig::First => HiddenFileSort::First,
-                    BrowserSortHiddenConfig::Last => HiddenFileSort::Last,
-                    BrowserSortHiddenConfig::Mixed => HiddenFileSort::Mixed,
-                })
-                .build()
-        });
-
-        let request = Request::builder()
-            .params(RequestParams::GetFiles(
-                GetFilesRequestParams::builder()
-                    .dir(args.dir().clone().unwrap_or_else(current_dir::current_dir))
-                    .sort(sort)
-                    .build(),
-            ))
-            .build();
+        let request = get_files_request(
+            args.dir().clone().unwrap_or_else(current_dir::current_dir),
+            &config,
+            config.browser().metadata(),
+        );
 
         let request_uuid: Uuid = *request.uuid();
         let effect = SystemEffect::Request(request);